+++ PokemonUnity.Shared/Framework/Essentials/IGameScreen.cs
@@ -43,7 +43,7 @@ namespace PokemonEssentials.Interface
 		/// </summary>
 		/// <param name="tone">color tone</param>
 		/// <param name="duration">time</param>
-		void start_tone_change(ITone tone, int duration);
+		void start_tone_change(ITone tone, float duration);
 		/// <summary>
 		/// Start Flashing
 		/// </summary>
